Generalize usage statements of dewdrop
This patch introduces a decorator which helps to specify which
subcommands correspond to which methods in ddr.py

Additionally the docstrings were updated to provide usage information
to the programmer & user at the same time.

The first line should be a summary of the usage of the method, and the
subsequent lines will be printed out during usage() statements.

This is an abuse of docstrings, but I felt they were more useful when
used in this manner than by giving myself notes.